refactor(reporting): Allow more control about test summary write process
This fixes a TODO regarding the hardcoded paths in
`AnvilTestExecutionListener`. This makes it possible to `TestSummary`
accessible programmatically and also makes it possible to write the JSON
report to arbitrary output streams.

The file is not written automatically when the test plan finishes
anymore. To restore the behavior as before, use the following code:

    launcher.execute(launcherDiscoveryRequest);

    try(OutputStream outputStream = Files.newOutputStream(Paths.get("resources/out", "summary.json"))) {
        executionListener.writeTestSummary(outputStream);
    } catch(IOException e) {
        e.printStackTrace();
    }