refactor(model): Remove unnecessary `ConfigContainer` class
This simplifies the library by getting rid of the unnecessary
`ConfigContainer` type. Neither TLS-Anvil, nor the current state of
x509-Anvil or SSH-Anvil need support for multiple independent
Configuration classes. It's usually not wise to add code that is
actually not needed under the presumption that it might be needed later,
because it's very likely that ["you ain't gonna need it"][YAGNI].

Anvil-Core is still under development and hasn't settled on a final API,
so we could easily add such a feature if deemed necessary, but it would
also be possible to simply combine those different configurations into a
single `AnvilConfig` implementation without the extra complexity:

    public class CompoundConfig implements AnvilConfig {
        protected final SomeConfig someConfig = new SomeConfig();
        protected final AnotherConfig anotherConfig = new AnotherConfig();

        public SomeConfig getSomeConfig() {
            return this.someConfig;
        }

        public AnotherConfig getAnotherConfig() {
            return this.anotherConfig;
        }
    }

[YAGNI]: https://www.martinfowler.com/bliki/Yagni.html